+++ __test__/expected/issue_10.js
@@ -1,0 +1,3 @@
+function foo (info) {
+  info.push({ ...profiles.splice(pos, 1)[0].output, index });
+}