+++ __test__/fixtures/issue_10.js
@@ -1,0 +1,6 @@
+/*
+  Errors out on spread operator
+*/
+function foo (info) {
+  info.push({ ...profiles.splice(pos, 1)[0].output, index });
+}