+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # rollup-plugin-cleanup changes
 
+### v3.0.0
+- Removed the [riot](http://riotjs.com/) .tag extension from the defaults, you can add this manually.
+- Now acorn `ecmaVersion` defaults to 9 (ES2018).
+- Closes #10 : Errors out on spread operator.
+- Updated devDependencies the to last acorn and rollup plugins.
+
 ### v2.0.1
 - Adds rollup >=0.50 as peerDependencies, hope to update devDependencies soon\*.
 - Preserves empty lines inside multi-line strings.